new main file from scratch (saving old one for now)
